fix timestamps once and for all

+++ skylab/skylab.go
@@ -86,7 +86,7 @@ func ToCanFrame(p Packet) (id uint32, data []byte, err error) {
 // internal structure for partially decoding json object.
 // includes
 type RawJsonEvent struct {
-	Timestamp uint32          `json:"ts" db:"ts"`
+	Timestamp int64          `json:"ts" db:"ts"`
 	Id        uint32          `json:"id"`
 	Name      string          `json:"name"`
 	Data      json.RawMessage `json:"data"`
@@ -102,8 +102,8 @@ type BusEvent struct {
 func (e *BusEvent) MarshalJSON() (b []byte, err error) {
 	// create the underlying raw event
 	j := &RawJsonEvent{
-		Timestamp: uint32(e.Timestamp.UnixMilli()),
-		Id:        uint32(e.Id),
+		Timestamp: e.Timestamp.UnixMilli(),
+		Id:        e.Id,
 		Name:      e.Data.String(),
 	}
 	// now we use the magic Packet -> map[string]interface{} function
@@ -125,7 +125,7 @@ func (e *BusEvent) UnmarshalJSON(b []byte) error {
 		return err
 	}
 
-	e.Timestamp = time.UnixMilli(int64(j.Timestamp))
+	e.Timestamp = time.UnixMilli(j.Timestamp)
 	e.Id = j.Id
 	e.Data, err = FromJson(j.Id, j.Data)
